Move contents of `action-utils.ts` into `manifests-singleton.ts`
This avoids the circular dependency.
